ADD Tree node can use double cliked to open the tabText

+++ src/tw/Max/Class/FileTree.java
@@ -47,19 +47,21 @@ public class FileTree extends JTree {
 		Boolean backupStatus = sqlInsert.insertTabTextBackup(account, tabName);
 		
 		// 成功備份才能刪除檔案
-		if (backupStatus) {
-			// 刪除檔案
-			Boolean deleteStatus = sqlDelete.deleteTabText(account, tabName);
-			
-			// 刪除成功才能刪除Tree Node
-			if (deleteStatus) {
-				// 刪除Tree Node
-				root.remove(index);
-				model = new DefaultTreeModel(root);
-				setModel(model);
-				JOptionPane.showMessageDialog(null, "刪除成功");
-			} else {
-				JOptionPane.showMessageDialog(null, "刪除失敗");
+		if (isDelete()) {
+			if (backupStatus) {
+				// 刪除檔案
+				Boolean deleteStatus = sqlDelete.deleteTabText(account, tabName);
+				
+				// 刪除成功才能刪除Tree Node
+				if (deleteStatus) {
+					// 刪除Tree Node
+					root.remove(index);
+					model = new DefaultTreeModel(root);
+					setModel(model);
+					JOptionPane.showMessageDialog(null, "刪除成功");
+				} else {
+					JOptionPane.showMessageDialog(null, "刪除失敗");
+				}
 			}
 		}
 	}
